Changed the BowlingGameFun to comply to the instructions
Write a class named “Game” that has two methods
roll(pins : int) is called each time the player rolls a ball.  The
argument is the number of pins knocked down.
score() : int is called only at the very end of the game.  It returns
the total score for that game.
It is no longer pure functional, it keeps state: rolls is mutable and
roll(pins) has the side effect to update rolls